Fix serial char printing with delays
Read the busy flag after E is set to 0; this was mistakenly thought to be non-functional.
In the E pulse forming subroutine, reset all control lines to zero, not just the E pin. There is a strange artifact when printing characters consecutively but with significant delays between them (printing strings works fine). The issue arises when the last instruction before the next character printing was EPULSEDAT, which left the RS bit set. This caused WAITBUSY in the next printing instruction to ignore RW (even though it is set upon entry) and, instead of reading the busy flag, it writes an empty character, creating a gap on the screen. It seems that WAITBUSY ignores RW and assumes RS is still set (despite RS being cleared and RW being set with E in one instruction upon entry), leading to a write operation instead of a read.
Do minor refactor.